move: 폴더 생성, 이동, 정리
