+++ src/test/java/domain/CarNameTest.java
@@ -1,4 +1,4 @@
-package racinggame;
+package domain;
 
 import static org.assertj.core.api.Assertions.*;
