+++ src/test/java/domain/GameTryCountTest.java
@@ -1,4 +1,4 @@
-package racinggame;
+package domain;
 
 import static org.assertj.core.api.Assertions.*;
